23.05.11 update
Updated many things. I wont summ it up.
TODO:
- add img to products
- remove profile and order related stuff
- clean up the fucking routes

+++ ecom/database/migrations/2023_04_15_095439_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name', 255);
-            $table->integer('price')->default(0);
+            $table->bigInteger('price')->default(0);
             $table->softDeletes();
             $table->timestamps();
         });
